server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
chromadb
python-multipart
pymupdf
requests